+++ src/main/java/org/truffleruby/language/arguments/ReadCallerFrameNode.java
@@ -9,12 +9,13 @@
  */
 package org.truffleruby.language.arguments;
 
-import org.truffleruby.builtins.CallerFrameAccess;
+import com.oracle.truffle.api.CompilerDirectives;
+import com.oracle.truffle.api.frame.FrameInstance;
 import org.truffleruby.language.RubyBaseNode;
 import org.truffleruby.language.dispatch.CachedDispatchNode;
 
+import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
-import com.oracle.truffle.api.frame.Frame;
 import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.DirectCallNode;
@@ -24,18 +25,21 @@ import com.oracle.truffle.api.profiles.ConditionProfile;
 public class ReadCallerFrameNode extends RubyBaseNode {
 
     private final ConditionProfile callerFrameProfile = ConditionProfile.createBinaryProfile();
-
-    private final CallerFrameAccess accessMode;
+    @CompilationFinal private volatile boolean firstCall = true;
 
     public static ReadCallerFrameNode create() {
-        return new ReadCallerFrameNode(CallerFrameAccess.MATERIALIZE);
+        return new ReadCallerFrameNode();
     }
 
-    public ReadCallerFrameNode(CallerFrameAccess callerFrameAccess) {
-        this.accessMode = callerFrameAccess;
-    }
+    public MaterializedFrame execute(VirtualFrame frame) {
+        // Avoid polluting the profile for the first call which has to use getCallerFrame()
+        if (firstCall) {
+            CompilerDirectives.transferToInterpreterAndInvalidate();
+            firstCall = false;
+            notifyCallerToSendFrame();
+            return getCallerFrame();
+        }
 
-    public Frame execute(VirtualFrame frame) {
         final MaterializedFrame callerFrame = RubyArguments.getCallerFrame(frame);
 
         if (callerFrameProfile.profile(callerFrame != null)) {
@@ -45,7 +49,7 @@ public class ReadCallerFrameNode extends RubyBaseNode {
         }
     }
 
-    private void replaceDispatchNode() {
+    private void notifyCallerToSendFrame() {
         final Node callerNode = getContext().getCallStack().getCallerNode(0, false);
         if (callerNode instanceof DirectCallNode) {
             final Node parent = callerNode.getParent();
@@ -56,9 +60,8 @@ public class ReadCallerFrameNode extends RubyBaseNode {
     }
 
     @TruffleBoundary
-    private Frame getCallerFrame() {
-        replaceDispatchNode();
-        return getContext().getCallStack().getCallerFrameIgnoringSend().getFrame(accessMode.getFrameAccess());
+    private MaterializedFrame getCallerFrame() {
+        return getContext().getCallStack().getCallerFrameIgnoringSend().getFrame(FrameInstance.FrameAccess.MATERIALIZE).materialize();
     }
 
 }